Change generated unit test random seed to be explicitly unsigned int

+++ test/test_all.cpp
@@ -65,7 +65,7 @@ using namespace realm::test_util;
 using namespace realm::test_util::unit_test;
 
 // Random seed for various random number generators used by fuzzying unit tests.
-unsigned long unit_test_random_seed;
+unsigned int unit_test_random_seed;
 
 namespace {
